+++ bin/proving-service/src/commands/mod.rs
@@ -49,6 +49,9 @@ pub(crate) struct ProxyConfig {
     /// Supported proof types.
     #[clap(long, default_value = "tx,batch,block", env = "MPS_SUPPORTED_PROOF_TYPES")]
     pub(crate) supported_proof_types: String,
+    /// Status port.
+    #[clap(long, default_value = "8084", env = "MPS_STATUS_PORT")]
+    pub(crate) status_port: u16,
 }
 
 #[derive(Debug, Clone, clap::Parser)]